working

+++ src/components/Profile/ProfilePost.jsx
@@ -10,7 +10,8 @@ import useShowToast from "../../hooks/useShowToast";
 import { useState } from "react";
 import { deleteObject, ref } from "firebase/storage";
 import { firestore, storage } from "../../firebase/firebase";
-import { deleteDoc, doc } from "firebase/firestore";
+import { arrayRemove, deleteDoc, doc, updateDoc } from "firebase/firestore";
+import usePostStore from "../../store/postStore";
 
 
 const ProfilePost = ({ post }) => {
@@ -19,17 +20,27 @@ const ProfilePost = ({ post }) => {
   const authUser = useAuthStore((state) => state.user);
   const showToast = useShowToast();
   const [isDeleting, setIsDeleting] = useState(false);
+  const deletePost = usePostStore(state => state.deletePost);
+  const decrementPostsCount = useUserProfileStore((state) => state.deletePost);
   const handleDeletePost = async () => {
     if (!window.confirm("Are you sure you want to delete this post?")) return;
+    if (isDeleting) return;
     try {
       const imageRef = ref(storage, `posts/${post.id}`);
       await deleteObject(imageRef);
       const userRef = doc(firestore, "users", authUser.uid);
       await deleteDoc(doc(firestore, "posts", post.id));
+      await updateDoc(userRef, {
+        posts: arrayRemove(post.id)
+      })
+
+      deletePost(post.id);
+      decrementPostsCount(post.id);
+      showToast("Success", "Post deleted successfully", "success");
     } catch (error) {
       showToast("Error", error.message, 'error')
     } finally {
-
+      setIsDeleting(false)
     }
   }
 
@@ -81,7 +92,7 @@ const ProfilePost = ({ post }) => {
 
                   {authUser?.uid === userProfile.uid && (
                     <Button size={"sm"} bg={"transparent"} _hover={{ bg: "whiteAlpha.300", color: "red.600" }} borderRadius={4} p={1}
-                      onClick={handleDeletePost}>
+                      onClick={handleDeletePost} isLoading={isDeleting}>
                       <MdDelete size={20} cursor={"pointer"} />
                     </Button>
                   )}
